+++ loggers/winston/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+- Add support for [ECS tracing fields](https://www.elastic.co/guide/en/ecs/current/ecs-tracing.html).
+  If it is detected that [Elastic APM](https://www.npmjs.com/package/elastic-apm-node)
+  is in use and there is an active trace, then tracing fields will be added to
+  log records. This enables linking between traces and log records in Kibana.
+
 - Fix guarding of the top-level 'log' and 'log.level' fields. A log statement
   can now set ['log' fields](https://www.elastic.co/guide/en/ecs/current/ecs-log.html)
   e.g.: